Nemo/requirements.txt
# =============================================================================
# Nemo - Real-time AI Companion Backend
# Python Dependencies
# =============================================================================

# -----------------------------------------------------------------------------
# Web Framework & Server
# -----------------------------------------------------------------------------
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0

# -----------------------------------------------------------------------------
# LLM Client (Groq - Ultra-fast inference)
# -----------------------------------------------------------------------------
groq>=0.4.0

# -----------------------------------------------------------------------------
# Memory Service (Mem0)
# -----------------------------------------------------------------------------
mem0ai>=0.1.0

# -----------------------------------------------------------------------------
# Text-to-Speech (Edge-TTS - Free, no API key!)
# -----------------------------------------------------------------------------
edge-tts>=6.1.0

# -----------------------------------------------------------------------------
# Environment & Utilities
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0
httpx>=0.26.0

# -----------------------------------------------------------------------------
# Optional: Speech-to-Text (for future STT integration)
# -----------------------------------------------------------------------------
# deepgram-sdk>=3.0.0
